Handled edge case for deleting category. (#60)

+++ app/javascript/src/components/Dashboard/Settings/ManageCategories/DeleteModal.jsx
@@ -17,14 +17,21 @@ const DeleteModal = ({
   const handleSubmit = async id => {
     setShowDeleteModal(false);
     try {
-      await articlesApi.batchUpdate({
-        previous_category_id: id,
-        updated_category_id: moveArticlesToCategory.value,
-      });
-      await categoriesApi.destroy(id);
-      Toastr.success(
-        `Articles successfully moved to the category: ${moveArticlesToCategory.label}`
-      );
+      if (categoryList.length > 1) {
+        await articlesApi.batchUpdate({
+          previous_category_id: id,
+          updated_category_id: moveArticlesToCategory.value,
+        });
+        await categoriesApi.destroy(id);
+        Toastr.success(
+          `Articles successfully moved to the category: ${moveArticlesToCategory.label}`
+        );
+      } else {
+        await categoriesApi.update(id, {
+          name: "General",
+        });
+        Toastr.success("Articles successfully moved to General category");
+      }
       refetch();
     } catch (error) {
       logger.error(error);
@@ -43,31 +50,48 @@ const DeleteModal = ({
           You are permanently deleting the {category.name} category. This action
           cannot be undone. Are you sure you wish to continue?
         </Typography>
-        <Callout icon={Warning} style="danger">
-          <div>
-            Category <strong>{category.name}</strong> has&nbsp;
-            <strong>
-              {category.articles.length}
-              {category.articles.length > 1 ? " articles" : " article"}
-            </strong>
-            . Before this category can be deleted these articles needs to be
-            moved to another category.
-          </div>
-        </Callout>
-        <Select
-          required
-          label="Select a category to move these articles into"
-          name="category"
-          placeholder="Select Category"
-          strategy="fixed"
-          options={categoryList
-            .filter(categoryItem => categoryItem.id !== category.id)
-            .map(category => ({
-              label: category.name,
-              value: category.id,
-            }))}
-          onChange={e => setMoveArticlesToCategory(e)}
-        />
+        {categoryList.length > 1 && (
+          <Callout icon={Warning} style="danger">
+            <div>
+              Category <strong>{category.name}</strong> has&nbsp;
+              <strong>
+                {category.articles.length}
+                {category.articles.length > 1 ? " articles" : " article"}
+              </strong>
+              . Before this category can be deleted these articles needs to be
+              moved to another category.
+            </div>
+          </Callout>
+        )}
+        {categoryList.length === 1 && (
+          <Callout icon={Warning} style="danger">
+            <div>
+              Category <strong>{category.name}</strong> has&nbsp;
+              <strong>
+                {category.articles.length}
+                {category.articles.length > 1 ? " articles" : " article"}
+              </strong>
+              . This will be moved to category general. Click proceed to
+              continue.
+            </div>
+          </Callout>
+        )}
+        {categoryList.length > 1 && (
+          <Select
+            required
+            label="Select a category to move these articles into"
+            name="category"
+            placeholder="Select Category"
+            strategy="fixed"
+            options={categoryList
+              .filter(categoryItem => categoryItem.id !== category.id)
+              .map(category => ({
+                label: category.name,
+                value: category.id,
+              }))}
+            onChange={e => setMoveArticlesToCategory(e)}
+          />
+        )}
       </Modal.Body>
       <Modal.Footer>
         <Button
